Actualización App
Se refactoriza la View Home,  se crea la CardViewTribute, CardViewCarousel y se mejora el estado de las tarjetas, se implementa un botón el la parte superior izquierda para llamar a la vista "CardViewTribute", se le añade mas detalles en la vista " DetailsView", se eliminan los comentarios y códigos que no son necesario, y se crea una carpeta llamada cajón desastre, donde se guarda vista y fragmento de código, que podría se reutilizado durante el desarrollo
