+++ DrgonBollWiki/Cajon_Desastre/ParallaxHeder.swift
@@ -1,0 +1,30 @@
+//
+//  ParallaxHeder.swift
+//  DrgonBollWiki
+//
+//  Created by Grupo de Estudio on 1/3/24.
+//
+
+//import SwiftUI
+//
+//struct ParallaxHeder<Content:View>: View {
+//    
+//    @ViewBuilder var content: () -> Content
+//    
+//    var body: some View {
+//        GeometryReader { geometry in
+//            let minX = geometry.frame(in: .global).minX
+//            
+//            content()
+//                .frame(
+//                    width: geometry.size.width,
+//                    height: geometry.size.height + max(minX, 10)
+//                )
+//                .offset(x:  +minX * 0.2)
+//              
+//        }
+//        
+//        
+//    }
+//}
+
