+++ DrgonBollWiki/View/ DetailsView.swift
@@ -15,14 +15,26 @@ struct DetailsView: View {
         _singleCharacterViewModel = State(wrappedValue: SingleCharacterViewModel(singleCharacterDataService: singleCharactersDataService))
         self.characterId = characterId
     }
-    
    
     
-   // @Binding var characterId: Int?
-     
     var body: some View {
-        Text(singleCharacterViewModel.character?.name ?? "").task {
-            await singleCharacterViewModel.getCharacterInformation(characterID: characterId )
+        ScrollView{
+            VStack(alignment: .leading){
+              
+                Text(singleCharacterViewModel.character?.name ?? "").task {
+                    await singleCharacterViewModel.getCharacterInformation(characterID: characterId )
+                }.font(.custom("Saiyan.ttf", size: 40)).padding(.horizontal, 100).bold()
+                Divider()
+                Text(singleCharacterViewModel.character?.description ?? "").task {
+                    await singleCharacterViewModel.getCharacterInformation(characterID: characterId )
+                }
+                Divider()
+                Text(singleCharacterViewModel.character?.originPlanet.description ?? "").task {
+                    await singleCharacterViewModel.getCharacterInformation(characterID: characterId )
+                }
+               
+            }.padding()
+          
         }
     }
    
@@ -36,29 +48,3 @@ extension DetailsView {
 
 }
 
-/*
- struct DetailsView: View {
-     @State private var singleCharacterViewModel: SingleCharacterViewModel
-      var characterId: Int?
-     
-     init(singleCharactersDataService: SingleCharacterProtocol) {
-         _singleCharacterViewModel = State(wrappedValue: SingleCharacterViewModel(singleCharacterDataService: singleCharactersDataService))
-         
-         }
-     
-    // @Binding var characterId: Int?
-      
-     var body: some View {
-         Text(singleCharacterViewModel.character?.name ?? "").task {
-             if let id = characterId {
-                 await singleCharacterViewModel.getCharacterInformation(characterID: id )
-             }
-             
-         }
-     }
- }
-
- #Preview {
-     DetailsView(singleCharactersDataService: MockSingleCharacterDataService(testData: nil))
- }
- */
