+++ DrgonBollWiki/View/Home.swift
@@ -8,33 +8,27 @@
 import SwiftUI
 
 struct Home: View {
+    
+    // MARK: - Se intacia SwiftData
     @State var dbSwiftDataModel: [DbSwiftDataModel]
     @State var dbSwiftDataViewModel: DbSwiftDataViewModel
+    
     @State private var homeViewModel: HomeViewModel
     @State private var planetsViewModel: PlanetsViewModel
     @State private var selectedCharacter: Character?
     @State private var selectedCharacterId: Int?
     
-    @State private var isFlipped = false
-    @State private var isShowDetails = false
-    @State private var favoritesStar = false
-    @State private var modeViewCard = false
-    @State private var isProgress = 0.6
     
+    @State private var favoritesStar = false
+    @State private var isTribute = false
+    
+    // MARK: - Numero de incice para el cambio de vistas
     @State var isShow: Int = 2
     @State var isShow2: Int = 1
     
     @State private var offset: CGFloat = 0
     @State private var scrollSpeed: CGFloat = 0
     
-    
-    /// View Properties
-    @State private var isRotationEnabled: Bool = true
-    @State private var showIndicator: Bool = false
-    
-    //@Environment (dbSwiftDataViewModel.self) var viewModelisFavorites
-    
-    let color: [Color] = [.red, .blue, .cyan, .yellow]
     let timer = Timer.publish(every: 0.02, on: .main, in: .common).autoconnect()
     
     
@@ -58,161 +52,22 @@ struct Home: View {
                 
                     .frame(width: 600, height: 800)
                     .opacity(0.6)
-                if liveScrol(isShow2) {
+                
+                if homeViewModel.liveScrol(isShow) {
                     
-                    if homeViewModel.isLoading{
-                        ProgressView()
-                    }else if let characters = homeViewModel.allCharacters {
-                        NavigationStack{
-                            VStack{
-                                GeometryReader{ let size = $0.size
-                                    ScrollView(.horizontal){
-                                        HStack(spacing: 0){
-                                            ForEach(characters.items, id:\.id){ character in
-                                                ZStack{
-                                                    ColorView(color: color[character.id % color.count]).padding(.bottom, 90)
-                                                    // CardView( item, character)
-                                                    Card(character: character, playSound: {
-                                                        homeViewModel.playCardSound() }).shadow(radius: /*@START_MENU_TOKEN@*/10/*@END_MENU_TOKEN@*/)
-                                                }                                        .contextMenu(ContextMenu(menuItems: {
-                                                    Button(action: {
-                                                        selectedCharacterId = character.id
-                                                        isShowDetails = true
-                                                    }, label: {
-                                                        Text("Saber Mas")
-                                                        Image(systemName: "book")
-                                                    })
-                                                    Button(action: {
-                                                        //                                                favoritesStar = true
-                                                        //                                                viewModelisFavorites.saveFavorites(character.id, favoritesStar)
-                                                    }, label: {
-                                                        Text("Guardar en favoritos")
-                                                        Image(systemName: "star.fill")
-                                                    })
-                                                    
-                                                    Button(action: {
-                                                        // logica
-                                                    }, label: {
-                                                        Text("Compartir")
-                                                        Image(systemName: "square.and.arrow.up")
-                                                    })
-                                                    
-                                                    Button(action: {
-                                                        // logica
-                                                    }, label: {
-                                                        Text("Copiar")
-                                                        Image(systemName: "doc.on.doc")
-                                                    })
-                                                    
-                                                }))
-                                                .sheet(isPresented: $isShowDetails, content: {
-                                                    DetailsView(singleCharactersDataService: SingleCharacterDataService(), characterId: selectedCharacterId ?? 1)
-                                                })
-                                                .padding(.horizontal, 65)
-                                                .frame(width: size.width)
-                                                .visualEffect { content, geometryProxy in
-                                                    content
-                                                        .scaleEffect(scale(geometryProxy,scale: 0.1), anchor: .trailing)
-                                                        .rotationEffect(rotation(geometryProxy, rotation: isRotationEnabled ? 5 : 0))
-                                                        .offset(x: minX(geometryProxy))
-                                                        .offset(x: excessMinX(geometryProxy, offset: isRotationEnabled ? 8 : 10))
-                                                    
-                                                }
-                                                .zIndex(characters.items.zIndex(character))
-                                                
-                                            }
-                                        }
-                                        .padding(.vertical, 15)
-                                        
-                                        
-                                    }
-                                    .scrollTargetBehavior(.paging)
-                                    .scrollIndicators(showIndicator ? .visible : .hidden)
-                                    .scrollIndicatorsFlash(trigger: showIndicator)
-                                }
-                                .frame(height: 410)
-                                .animation(.snappy, value: isRotationEnabled)
-                                if modeViewCard {
-                                    VStack(spacing: 10){
-                                        Toggle("Rotation Enable", isOn: $isRotationEnabled)
-                                        Toggle("Shows Scrol Indiquetor", isOn: $showIndicator)
-                                    }
-                                    .padding(15)
-                                    .background(.bar, in: .rect(cornerRadius: 10))
-                                    .padding(15)
-                                    
-                                }
-                            }.padding(100)
-                        }
-                    }
-                }else if liveScrol(isShow) {
-                    if homeViewModel.isLoading{
-                        ProgressView()
-                    }else if let characters = homeViewModel.allCharacters {
-                        ScrollView(.horizontal){
-                            HStack{
-                                ForEach(characters.items, id:\.id){ character in
-                                    ZStack{
-                                        ColorView(color: color[character.id % color.count]).padding(.bottom, 90).opacity(0.2)
-                                        Card(character: character, playSound: {
-                                            homeViewModel.playCardSound() })
-                                        .padding()
-                                        .onTapGesture {
-                                            withAnimation {
-                                                isFlipped.toggle()
-                                            }
-                                        }
-                                    }
-                                    .contextMenu(ContextMenu(menuItems: {
-                                        Button(action: {
-                                            selectedCharacterId = character.id
-                                            isShowDetails = true
-                                        }, label: {
-                                            Text("Saber Mas")
-                                            Image(systemName: "book")
-                                        })
-                                        Button(action: {
-                                            //                                                favoritesStar = true
-                                            //                                                viewModelisFavorites.saveFavorites(character.id, favoritesStar)
-                                        }, label: {
-                                            Text("Guardar en favoritos")
-                                            Image(systemName: "star.fill")
-                                        })
-                                        
-                                        Button(action: {
-                                            // logica
-                                        }, label: {
-                                            Text("Compartir")
-                                            Image(systemName: "square.and.arrow.up")
-                                        })
-                                        
-                                        Button(action: {
-                                            // logica
-                                        }, label: {
-                                            Text("Copiar")
-                                            Image(systemName: "doc.on.doc")
-                                        })
-                                        
-                                    }))
-                                    .sheet(isPresented: $isShowDetails, content: {
-                                        DetailsView(singleCharactersDataService: SingleCharacterDataService(), characterId: selectedCharacterId ?? 1)
-                                    })
-                                }
-                            }.scrollTargetLayout()
-                                .padding(.horizontal)
-                        } .scrollTargetBehavior(.viewAligned)
-                            .padding(.horizontal, 100)
-                    }else {
-                        VStack{
-                            ProgressView(value: isProgress).progressViewStyle(.circular)
-                            Text("Please wait ...")
-                                .font(.title).bold()
-                                .foregroundStyle(Color.white).shadow(radius: 10)
-                        }
-                        
-                    }
+                    // MARK: - View the CardView
+                    CardView(allCaractersDataService: AllCharactersDataService(),
+                             dbSwiftDataModel: [],
+                             dbSwiftDataViewModel: DbSwiftDataViewModel()).padding(.horizontal ,105)
                     
                     
+                }else if homeViewModel.liveScrol(isShow2) {
+                    
+                    // MARK: - View the CardViewCarousel
+                    CardViewCarousel(allCaractersDataService: AllCharactersDataService(),
+                                     planetsDataSevice: PlanetsDataService(),
+                                     dbSwiftDataModel: [],
+                                     dbSwiftDataViewModel: DbSwiftDataViewModel())
                 }
                 
                 VStack{
@@ -220,7 +75,7 @@ struct Home: View {
                     HStack(alignment: .top){
                         Image("Boll7")
                             .resizable()
-                            .frame(width: 500, height: 500)
+                            .frame(width: 495, height: 495)
                             .padding(.leading, -130)
                             .padding(.bottom, -300)
                             .shadow(radius: 8)
@@ -235,30 +90,31 @@ struct Home: View {
                             isShow = 2
                             isShow2 = 1
                         }) {
-                            Text("Vista de Cartas")
+                          //  Image(systemName: (isShow == 1) ? "" : "checkmark")
+                            Text("Vista de carrusel")
                             Image( "Boll7")
                                 .resizable()
                                 .frame(width: 10, height: 10)
-                        }
+                        }.disabled((((isShow == 2) ? 2 : 1) == 2))
                         
                         Button(action: {
                             isShow = 1
                             isShow2 = 2
                         }) {
-                            Text("Vista de carrusel")
+                           // Image(systemName: (isShow == 2) ? "" : "checkmark")
+                            Text("Vista de cartas")
                             
                             Image( "Dragon")
                                 .resizable()
                                 .frame(width: 10, height: 10)
-                        }
+                        }.disabled((((isShow == 1) ? 1 : 2) == 1))
                         
                         Button(action: {
-                            // Acción del botón de la barra de herramientas
-                            modeViewCard.toggle()
+                            //Acer algo
                         }) {
                             
                             //  Image(systemName: "person")
-                            Text("Comfigurar vista de cartas")
+                            Text("Items")
                             Image( "Mutenroy")
                                 .resizable()
                                 .frame(width: 10, height: 10)
@@ -278,59 +134,29 @@ struct Home: View {
                         Image("logoGoku").resizable()
                             .frame(width: 40, height: 50)
                     }
-                    
-                    
-                    
+                }
+                ToolbarItem(placement: .navigation) {
+                    Menu{
+                        
+                        Button(action: {
+                            // Acción del botón de la barra de herramientas
+                            isTribute = true
+                        }) {
+                            Text("Tributo a Akira Toriyama")
+                            Image( "akira")
+                                .resizable()
+                                .frame(width: 10, height: 10)
+                        }
+                        
+                    } label: {
+                        Image(systemName: "info.circle").foregroundStyle(Color.black).bold().font(.custom("", size: 20))
+                    }
                 }
             }
-            
+            .sheet(isPresented: $isTribute, content: {
+                CardViewTribute()
+            })
         }
-    }
-    
-    func liveScrol(_ index: Int) -> Bool {
-        switch index {
-        case 1:
-            return true
-        case 2:
-            return false
-        default:
-            return false
-        }
-    }
-    
-    /// Stacked Card Animation
-    func minX(_ proxy: GeometryProxy) -> CGFloat {
-        let minX = proxy.frame(in: .scrollView(axis: .horizontal)).minX
-        return minX < 0 ? 0: -minX
-    }
-    
-    func progress(_ proxy: GeometryProxy, limit: CGFloat = 2) -> CGFloat {
-        let maxX = proxy.frame(in: .scrollView(axis: .horizontal)).maxX
-        let width = proxy.bounds(of: .scrollView(axis: .horizontal))?.width ?? 0
-        
-        /// Converting into Progress
-        let progress = (maxX / width) - 1.0
-        let cappedProgress = min(progress, limit)
-        
-        return cappedProgress
-    }
-    
-    func scale(_ proxy: GeometryProxy, scale: CGFloat = 0.1) -> CGFloat {
-        let progress = progress(proxy, limit: 3)
-        
-        return 1 - (progress * scale)
-    }
-    
-    func excessMinX(_ proxy: GeometryProxy, offset: CGFloat = 10) -> CGFloat {
-        let process = progress(proxy)
-        
-        return process * offset
-    }
-    
-    func rotation(_ proxy: GeometryProxy, rotation: CGFloat = 5) -> Angle {
-        let process = progress(proxy)
-        
-        return .init(degrees: process * rotation)
     }
 }
